Refactor thermal module and update matrix assembly routines
- Renamed `Solver_Time` module to `Time_Time` for clarity.
- Commented out unused variables and procedures in `Main_FTDSS` and `Main_Thermal`.
- Updated `Type_Output` initialization in `Main_FTDSS` to use the correct parameters.
- Modified `Assemble_Mass_Heat_1` and `Assemble_Diffusion_Heat_1` to include porosity and temperature in calculations.
- Adjusted `Material_Manager` to allocate additional properties for unique regions.
- Cleaned up debug print statements and commented out unused code in various modules.
- Updated test program to reflect changes in thermal domain and boundary conditions.

+++ src/Core/VTK_Constants.F90
@@ -1,5 +1,6 @@
 
 module Core_VTK_Constants
+    use, intrinsic :: iso_fortran_env, only: int32
     implicit none
     private
     public :: VTK_VERTEX, VTK_LINE, VTK_TRIANGLE, VTK_TETRA, VTK_QUAD, &
@@ -7,19 +8,19 @@ module Core_VTK_Constants
               VTK_QUADRATIC_TETRA, VTK_QUADRATIC_HEXAHEDRON, VTK_WEDGE, &
               VTK_PYRAMID, VTK_PIXEL, VTK_HEXAHEDRON
 
-    integer, parameter :: VTK_VERTEX = 1
-    integer, parameter :: VTK_LINE = 3
-    integer, parameter :: VTK_TRIANGLE = 5
-    integer, parameter :: VTK_PIXEL = 8
-    integer, parameter :: VTK_QUAD = 9
-    integer, parameter :: VTK_TETRA = 10
-    integer, parameter :: VTK_HEXAHEDRON = 12
-    integer, parameter :: VTK_WEDGE = 13
-    integer, parameter :: VTK_PYRAMID = 14
-    integer, parameter :: VTK_QUADRATIC_EDGE = 21
-    integer, parameter :: VTK_QUADRATIC_TRIANGLE = 22
-    integer, parameter :: VTK_QUADRATIC_QUAD = 23
-    integer, parameter :: VTK_QUADRATIC_TETRA = 24
-    integer, parameter :: VTK_QUADRATIC_HEXAHEDRON = 25
+    integer(int32), parameter :: VTK_VERTEX = 1
+    integer(int32), parameter :: VTK_LINE = 3
+    integer(int32), parameter :: VTK_TRIANGLE = 5
+    integer(int32), parameter :: VTK_PIXEL = 8
+    integer(int32), parameter :: VTK_QUAD = 9
+    integer(int32), parameter :: VTK_TETRA = 10
+    integer(int32), parameter :: VTK_HEXAHEDRON = 12
+    integer(int32), parameter :: VTK_WEDGE = 13
+    integer(int32), parameter :: VTK_PYRAMID = 14
+    integer(int32), parameter :: VTK_QUADRATIC_EDGE = 21
+    integer(int32), parameter :: VTK_QUADRATIC_TRIANGLE = 22
+    integer(int32), parameter :: VTK_QUADRATIC_QUAD = 23
+    integer(int32), parameter :: VTK_QUADRATIC_TETRA = 24
+    integer(int32), parameter :: VTK_QUADRATIC_HEXAHEDRON = 25
 
 end module Core_VTK_Constants
